+++ apps/zipsync/src/cli/test/start.test.ts
