+++ apps/zipsync/src/start.ts
@@ -5,7 +5,7 @@ import { ConsoleTerminalProvider } from '@rushstack/terminal/lib/ConsoleTerminal
 import { Terminal } from '@rushstack/terminal/lib/Terminal';
 
 import { version } from '../package.json';
-import { ZipSyncCommandLineParser } from './ZipSyncCommandLineParser';
+import { ZipSyncCommandLineParser } from './cli/ZipSyncCommandLineParser';
 
 const toolVersion: string = version;
