+++ apps/zipsync/src/test/testUtils.ts
@@ -5,7 +5,7 @@ import { tmpdir } from 'node:os';
 import * as path from 'node:path';
 import * as fs from 'node:fs';
 import * as crypto from 'node:crypto';
-import type { IMetadata } from './zipSyncUtils';
+import type { IMetadata } from '../zipSyncUtils';
 
 export function getTempDir(): string {
   const randomId: string = crypto.randomUUID();
